changed

+++ python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,33 +1,22 @@
 #!/usr/bin/python3
-"""
-Deletes all State objects with a name containing the letter a
-from the database hbtn_0e_6_usa.
-"""
+# Deletes all State objects with a name containing
+# the letter a from the database hbtn_0e_6_usa.
+# Usage: ./13-model_state_delete_a.py <mysql username> /
+#                                     <mysql password> /
+#                                     <database name>
 import sys
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from model_state import Base, State
+from model_state import State
 
 if __name__ == "__main__":
-    username = sys.argv[1]
-    password = sys.argv[2]
-    db_name = sys.argv[3]
-
-    engine = create_engine(
-        'mysql+mysqldb://{}:{}@localhost:3306/{}'.format(
-            username, password, db_name),
-        pool_pre_ping=True
-    )
-
+    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
+                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
+                           pool_pre_ping=True)
     Session = sessionmaker(bind=engine)
     session = Session()
 
-    to_delete = session.query(State).filter(
-        State.name.like('%a%')
-    ).all()
-
-    for state in to_delete:
-        session.delete(state)
-
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
     session.commit()
-    session.close()
